+++ apps/server/src/persistence/AuthSessions.ts
@@ -1,27 +1,109 @@
-import { AuthEnvironmentScopes, AuthSessionId, ServerAuthSessionMethod } from "@t3tools/contracts";
-import * as SqlClient from "effect/unstable/sql/SqlClient";
-import * as SqlSchema from "effect/unstable/sql/SqlSchema";
+import * as Context from "effect/Context";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 import * as Option from "effect/Option";
 import * as Schema from "effect/Schema";
+import * as SqlClient from "effect/unstable/sql/SqlClient";
+import * as SqlSchema from "effect/unstable/sql/SqlSchema";
 
 import {
+  AuthClientMetadataDeviceType,
+  AuthEnvironmentScopes,
+  AuthSessionId,
+  ServerAuthSessionMethod,
+} from "@t3tools/contracts";
+
+import {
+  type AuthSessionRepositoryError,
   toPersistenceDecodeError,
   toPersistenceSqlError,
-  type AuthSessionRepositoryError,
-} from "../Errors.ts";
-import {
-  AuthSessionRecord,
+} from "./Errors.ts";
+
+export const AuthSessionClientMetadataRecord = Schema.Struct({
+  label: Schema.NullOr(Schema.String),
+  ipAddress: Schema.NullOr(Schema.String),
+  userAgent: Schema.NullOr(Schema.String),
+  deviceType: AuthClientMetadataDeviceType,
+  os: Schema.NullOr(Schema.String),
+  browser: Schema.NullOr(Schema.String),
+});
+export type AuthSessionClientMetadataRecord = typeof AuthSessionClientMetadataRecord.Type;
+
+export const AuthSessionRecord = Schema.Struct({
+  sessionId: AuthSessionId,
+  subject: Schema.String,
+  scopes: AuthEnvironmentScopes,
+  method: ServerAuthSessionMethod,
+  client: AuthSessionClientMetadataRecord,
+  issuedAt: Schema.DateTimeUtcFromString,
+  expiresAt: Schema.DateTimeUtcFromString,
+  lastConnectedAt: Schema.NullOr(Schema.DateTimeUtcFromString),
+  revokedAt: Schema.NullOr(Schema.DateTimeUtcFromString),
+});
+export type AuthSessionRecord = typeof AuthSessionRecord.Type;
+
+export const CreateAuthSessionInput = Schema.Struct({
+  sessionId: AuthSessionId,
+  subject: Schema.String,
+  scopes: AuthEnvironmentScopes,
+  method: ServerAuthSessionMethod,
+  client: AuthSessionClientMetadataRecord,
+  issuedAt: Schema.DateTimeUtcFromString,
+  expiresAt: Schema.DateTimeUtcFromString,
+});
+export type CreateAuthSessionInput = typeof CreateAuthSessionInput.Type;
+
+export const GetAuthSessionByIdInput = Schema.Struct({
+  sessionId: AuthSessionId,
+});
+export type GetAuthSessionByIdInput = typeof GetAuthSessionByIdInput.Type;
+
+export const ListActiveAuthSessionsInput = Schema.Struct({
+  now: Schema.DateTimeUtcFromString,
+});
+export type ListActiveAuthSessionsInput = typeof ListActiveAuthSessionsInput.Type;
+
+export const RevokeAuthSessionInput = Schema.Struct({
+  sessionId: AuthSessionId,
+  revokedAt: Schema.DateTimeUtcFromString,
+});
+export type RevokeAuthSessionInput = typeof RevokeAuthSessionInput.Type;
+
+export const RevokeOtherAuthSessionsInput = Schema.Struct({
+  currentSessionId: AuthSessionId,
+  revokedAt: Schema.DateTimeUtcFromString,
+});
+export type RevokeOtherAuthSessionsInput = typeof RevokeOtherAuthSessionsInput.Type;
+
+export const SetAuthSessionLastConnectedAtInput = Schema.Struct({
+  sessionId: AuthSessionId,
+  lastConnectedAt: Schema.DateTimeUtcFromString,
+});
+export type SetAuthSessionLastConnectedAtInput = typeof SetAuthSessionLastConnectedAtInput.Type;
+
+export class AuthSessionRepository extends Context.Service<
   AuthSessionRepository,
-  type AuthSessionRepositoryShape,
-  CreateAuthSessionInput,
-  GetAuthSessionByIdInput,
-  ListActiveAuthSessionsInput,
-  RevokeAuthSessionInput,
-  RevokeOtherAuthSessionsInput,
-  SetAuthSessionLastConnectedAtInput,
-} from "../Services/AuthSessions.ts";
+  {
+    readonly create: (
+      input: CreateAuthSessionInput,
+    ) => Effect.Effect<void, AuthSessionRepositoryError>;
+    readonly getById: (
+      input: GetAuthSessionByIdInput,
+    ) => Effect.Effect<Option.Option<AuthSessionRecord>, AuthSessionRepositoryError>;
+    readonly listActive: (
+      input: ListActiveAuthSessionsInput,
+    ) => Effect.Effect<ReadonlyArray<AuthSessionRecord>, AuthSessionRepositoryError>;
+    readonly revoke: (
+      input: RevokeAuthSessionInput,
+    ) => Effect.Effect<boolean, AuthSessionRepositoryError>;
+    readonly revokeAllExcept: (
+      input: RevokeOtherAuthSessionsInput,
+    ) => Effect.Effect<ReadonlyArray<AuthSessionId>, AuthSessionRepositoryError>;
+    readonly setLastConnectedAt: (
+      input: SetAuthSessionLastConnectedAtInput,
+    ) => Effect.Effect<void, AuthSessionRepositoryError>;
+  }
+>()("t3/persistence/AuthSessions/AuthSessionRepository") {}
 
 const AuthSessionDbRow = Schema.Struct({
   sessionId: AuthSessionId,
@@ -68,7 +150,7 @@ function toPersistenceSqlOrDecodeError(sqlOperation: string, decodeOperation: st
       : toPersistenceSqlError(sqlOperation)(cause);
 }
 
-const makeAuthSessionRepository = Effect.gen(function* () {
+export const make = Effect.gen(function* () {
   const sql = yield* SqlClient.SqlClient;
 
   const createSessionRow = SqlSchema.void({
@@ -197,7 +279,7 @@ const makeAuthSessionRepository = Effect.gen(function* () {
       `,
   });
 
-  const create: AuthSessionRepositoryShape["create"] = (input) =>
+  const create: AuthSessionRepository["Service"]["create"] = (input) =>
     createSessionRow(input).pipe(
       Effect.mapError(
         toPersistenceSqlOrDecodeError(
@@ -207,7 +289,7 @@ const makeAuthSessionRepository = Effect.gen(function* () {
       ),
     );
 
-  const getById: AuthSessionRepositoryShape["getById"] = (input) =>
+  const getById: AuthSessionRepository["Service"]["getById"] = (input) =>
     getSessionRowById(input).pipe(
       Effect.mapError(
         toPersistenceSqlOrDecodeError(
@@ -223,7 +305,7 @@ const makeAuthSessionRepository = Effect.gen(function* () {
       ),
     );
 
-  const listActive: AuthSessionRepositoryShape["listActive"] = (input) =>
+  const listActive: AuthSessionRepository["Service"]["listActive"] = (input) =>
     listActiveSessionRows(input).pipe(
       Effect.mapError(
         toPersistenceSqlOrDecodeError(
@@ -234,7 +316,7 @@ const makeAuthSessionRepository = Effect.gen(function* () {
       Effect.flatMap((rows) => Effect.succeed(rows.map((row) => toAuthSessionRecord(row)))),
     );
 
-  const revoke: AuthSessionRepositoryShape["revoke"] = (input) =>
+  const revoke: AuthSessionRepository["Service"]["revoke"] = (input) =>
     revokeSessionRows(input).pipe(
       Effect.mapError(
         toPersistenceSqlOrDecodeError(
@@ -245,7 +327,7 @@ const makeAuthSessionRepository = Effect.gen(function* () {
       Effect.map((rows) => rows.length > 0),
     );
 
-  const revokeAllExcept: AuthSessionRepositoryShape["revokeAllExcept"] = (input) =>
+  const revokeAllExcept: AuthSessionRepository["Service"]["revokeAllExcept"] = (input) =>
     revokeOtherSessionRows(input).pipe(
       Effect.mapError(
         toPersistenceSqlOrDecodeError(
@@ -256,7 +338,7 @@ const makeAuthSessionRepository = Effect.gen(function* () {
       Effect.map((rows) => rows.map((row) => row.sessionId)),
     );
 
-  const setLastConnectedAt: AuthSessionRepositoryShape["setLastConnectedAt"] = (input) =>
+  const setLastConnectedAt: AuthSessionRepository["Service"]["setLastConnectedAt"] = (input) =>
     setLastConnectedAtRow(input).pipe(
       Effect.mapError(
         toPersistenceSqlOrDecodeError(
@@ -273,10 +355,7 @@ const makeAuthSessionRepository = Effect.gen(function* () {
     revoke,
     revokeAllExcept,
     setLastConnectedAt,
-  } satisfies AuthSessionRepositoryShape;
+  } satisfies AuthSessionRepository["Service"];
 });
 
-export const AuthSessionRepositoryLive = Layer.effect(
-  AuthSessionRepository,
-  makeAuthSessionRepository,
-);
+export const layer = Layer.effect(AuthSessionRepository, make);
