+++ apps/server/src/persistence/ProjectionProjects.ts
@@ -1,0 +1,206 @@
+import * as Context from "effect/Context";
+import * as Effect from "effect/Effect";
+import * as Layer from "effect/Layer";
+import * as Option from "effect/Option";
+import * as Schema from "effect/Schema";
+import * as Struct from "effect/Struct";
+import * as SqlClient from "effect/unstable/sql/SqlClient";
+import * as SqlSchema from "effect/unstable/sql/SqlSchema";
+
+import { IsoDateTime, ModelSelection, ProjectId, ProjectScript } from "@t3tools/contracts";
+
+import { type ProjectionRepositoryError, toPersistenceSqlError } from "./Errors.ts";
+
+/**
+ * ProjectionProjectRepository - Projection repository interface for projects.
+ *
+ * Owns persistence operations for project rows in the orchestration projection
+ * read model.
+ *
+ * @module ProjectionProjectRepository
+ */
+
+export const ProjectionProject = Schema.Struct({
+  projectId: ProjectId,
+  title: Schema.String,
+  workspaceRoot: Schema.String,
+  defaultModelSelection: Schema.NullOr(ModelSelection),
+  scripts: Schema.Array(ProjectScript),
+  createdAt: IsoDateTime,
+  updatedAt: IsoDateTime,
+  deletedAt: Schema.NullOr(IsoDateTime),
+});
+export type ProjectionProject = typeof ProjectionProject.Type;
+
+export const GetProjectionProjectInput = Schema.Struct({
+  projectId: ProjectId,
+});
+export type GetProjectionProjectInput = typeof GetProjectionProjectInput.Type;
+
+export const DeleteProjectionProjectInput = Schema.Struct({
+  projectId: ProjectId,
+});
+export type DeleteProjectionProjectInput = typeof DeleteProjectionProjectInput.Type;
+
+/**
+ * ProjectionProjectRepository - Service tag for project projection persistence.
+ */
+export class ProjectionProjectRepository extends Context.Service<
+  ProjectionProjectRepository,
+  {
+    /**
+     * Insert or replace a projected project row.
+     *
+     * Upserts by `projectId` and persists scripts through JSON encoding.
+     */
+    readonly upsert: (row: ProjectionProject) => Effect.Effect<void, ProjectionRepositoryError>;
+
+    /**
+     * Read a projected project row by id.
+     */
+    readonly getById: (
+      input: GetProjectionProjectInput,
+    ) => Effect.Effect<Option.Option<ProjectionProject>, ProjectionRepositoryError>;
+
+    /**
+     * List all projected project rows.
+     *
+     * Returned in deterministic creation order.
+     */
+    readonly listAll: () => Effect.Effect<
+      ReadonlyArray<ProjectionProject>,
+      ProjectionRepositoryError
+    >;
+
+    /**
+     * Soft-delete a projected project row by id.
+     */
+    readonly deleteById: (
+      input: DeleteProjectionProjectInput,
+    ) => Effect.Effect<void, ProjectionRepositoryError>;
+  }
+>()("t3/persistence/ProjectionProjects/ProjectionProjectRepository") {}
+
+const ProjectionProjectDbRow = ProjectionProject.mapFields(
+  Struct.assign({
+    defaultModelSelection: Schema.NullOr(Schema.fromJsonString(ModelSelection)),
+    scripts: Schema.fromJsonString(Schema.Array(ProjectScript)),
+  }),
+);
+type ProjectionProjectDbRow = typeof ProjectionProjectDbRow.Type;
+
+export const make = Effect.gen(function* () {
+  const sql = yield* SqlClient.SqlClient;
+
+  const upsertProjectionProjectRow = SqlSchema.void({
+    Request: ProjectionProject,
+    execute: (row) =>
+      sql`
+        INSERT INTO projection_projects (
+          project_id,
+          title,
+          workspace_root,
+          default_model_selection_json,
+          scripts_json,
+          created_at,
+          updated_at,
+          deleted_at
+        )
+        VALUES (
+          ${row.projectId},
+          ${row.title},
+          ${row.workspaceRoot},
+          ${row.defaultModelSelection !== null ? JSON.stringify(row.defaultModelSelection) : null},
+          ${JSON.stringify(row.scripts)},
+          ${row.createdAt},
+          ${row.updatedAt},
+          ${row.deletedAt}
+        )
+        ON CONFLICT (project_id)
+        DO UPDATE SET
+          title = excluded.title,
+          workspace_root = excluded.workspace_root,
+          default_model_selection_json = excluded.default_model_selection_json,
+          scripts_json = excluded.scripts_json,
+          created_at = excluded.created_at,
+          updated_at = excluded.updated_at,
+          deleted_at = excluded.deleted_at
+      `,
+  });
+
+  const getProjectionProjectRow = SqlSchema.findOneOption({
+    Request: GetProjectionProjectInput,
+    Result: ProjectionProjectDbRow,
+    execute: ({ projectId }) =>
+      sql`
+        SELECT
+          project_id AS "projectId",
+          title,
+          workspace_root AS "workspaceRoot",
+          default_model_selection_json AS "defaultModelSelection",
+          scripts_json AS "scripts",
+          created_at AS "createdAt",
+          updated_at AS "updatedAt",
+          deleted_at AS "deletedAt"
+        FROM projection_projects
+        WHERE project_id = ${projectId}
+      `,
+  });
+
+  const listProjectionProjectRows = SqlSchema.findAll({
+    Request: Schema.Void,
+    Result: ProjectionProjectDbRow,
+    execute: () =>
+      sql`
+        SELECT
+          project_id AS "projectId",
+          title,
+          workspace_root AS "workspaceRoot",
+          default_model_selection_json AS "defaultModelSelection",
+          scripts_json AS "scripts",
+          created_at AS "createdAt",
+          updated_at AS "updatedAt",
+          deleted_at AS "deletedAt"
+        FROM projection_projects
+        ORDER BY created_at ASC, project_id ASC
+      `,
+  });
+
+  const deleteProjectionProjectRow = SqlSchema.void({
+    Request: DeleteProjectionProjectInput,
+    execute: ({ projectId }) =>
+      sql`
+        DELETE FROM projection_projects
+        WHERE project_id = ${projectId}
+      `,
+  });
+
+  const upsert: ProjectionProjectRepository["Service"]["upsert"] = (row) =>
+    upsertProjectionProjectRow(row).pipe(
+      Effect.mapError(toPersistenceSqlError("ProjectionProjectRepository.upsert:query")),
+    );
+
+  const getById: ProjectionProjectRepository["Service"]["getById"] = (input) =>
+    getProjectionProjectRow(input).pipe(
+      Effect.mapError(toPersistenceSqlError("ProjectionProjectRepository.getById:query")),
+    );
+
+  const listAll: ProjectionProjectRepository["Service"]["listAll"] = () =>
+    listProjectionProjectRows().pipe(
+      Effect.mapError(toPersistenceSqlError("ProjectionProjectRepository.listAll:query")),
+    );
+
+  const deleteById: ProjectionProjectRepository["Service"]["deleteById"] = (input) =>
+    deleteProjectionProjectRow(input).pipe(
+      Effect.mapError(toPersistenceSqlError("ProjectionProjectRepository.deleteById:query")),
+    );
+
+  return {
+    upsert,
+    getById,
+    listAll,
+    deleteById,
+  } satisfies ProjectionProjectRepository["Service"];
+});
+
+export const layer = Layer.effect(ProjectionProjectRepository, make);
