+++ apps/server/src/persistence/ProjectionThreadProposedPlans.ts
@@ -1,0 +1,152 @@
+import * as Context from "effect/Context";
+import * as Effect from "effect/Effect";
+import * as Layer from "effect/Layer";
+import * as Schema from "effect/Schema";
+import * as SqlClient from "effect/unstable/sql/SqlClient";
+import * as SqlSchema from "effect/unstable/sql/SqlSchema";
+
+import {
+  IsoDateTime,
+  OrchestrationProposedPlanId,
+  ThreadId,
+  TrimmedNonEmptyString,
+  TurnId,
+} from "@t3tools/contracts";
+
+import { type ProjectionRepositoryError, toPersistenceSqlError } from "./Errors.ts";
+
+export const ProjectionThreadProposedPlan = Schema.Struct({
+  planId: OrchestrationProposedPlanId,
+  threadId: ThreadId,
+  turnId: Schema.NullOr(TurnId),
+  planMarkdown: TrimmedNonEmptyString,
+  implementedAt: Schema.NullOr(IsoDateTime),
+  implementationThreadId: Schema.NullOr(ThreadId),
+  createdAt: IsoDateTime,
+  updatedAt: IsoDateTime,
+});
+export type ProjectionThreadProposedPlan = typeof ProjectionThreadProposedPlan.Type;
+
+export const ListProjectionThreadProposedPlansInput = Schema.Struct({
+  threadId: ThreadId,
+});
+export type ListProjectionThreadProposedPlansInput =
+  typeof ListProjectionThreadProposedPlansInput.Type;
+
+export const DeleteProjectionThreadProposedPlansInput = Schema.Struct({
+  threadId: ThreadId,
+});
+export type DeleteProjectionThreadProposedPlansInput =
+  typeof DeleteProjectionThreadProposedPlansInput.Type;
+
+export class ProjectionThreadProposedPlanRepository extends Context.Service<
+  ProjectionThreadProposedPlanRepository,
+  {
+    readonly upsert: (
+      proposedPlan: ProjectionThreadProposedPlan,
+    ) => Effect.Effect<void, ProjectionRepositoryError>;
+    readonly listByThreadId: (
+      input: ListProjectionThreadProposedPlansInput,
+    ) => Effect.Effect<ReadonlyArray<ProjectionThreadProposedPlan>, ProjectionRepositoryError>;
+    readonly deleteByThreadId: (
+      input: DeleteProjectionThreadProposedPlansInput,
+    ) => Effect.Effect<void, ProjectionRepositoryError>;
+  }
+>()("t3/persistence/ProjectionThreadProposedPlans/ProjectionThreadProposedPlanRepository") {}
+
+export const make = Effect.gen(function* () {
+  const sql = yield* SqlClient.SqlClient;
+
+  const upsertProjectionThreadProposedPlanRow = SqlSchema.void({
+    Request: ProjectionThreadProposedPlan,
+    execute: (row) => sql`
+      INSERT INTO projection_thread_proposed_plans (
+        plan_id,
+        thread_id,
+        turn_id,
+        plan_markdown,
+        implemented_at,
+        implementation_thread_id,
+        created_at,
+        updated_at
+      )
+      VALUES (
+        ${row.planId},
+        ${row.threadId},
+        ${row.turnId},
+        ${row.planMarkdown},
+        ${row.implementedAt},
+        ${row.implementationThreadId},
+        ${row.createdAt},
+        ${row.updatedAt}
+      )
+      ON CONFLICT (plan_id)
+      DO UPDATE SET
+        thread_id = excluded.thread_id,
+        turn_id = excluded.turn_id,
+        plan_markdown = excluded.plan_markdown,
+        implemented_at = excluded.implemented_at,
+        implementation_thread_id = excluded.implementation_thread_id,
+        created_at = excluded.created_at,
+        updated_at = excluded.updated_at
+    `,
+  });
+
+  const listProjectionThreadProposedPlanRows = SqlSchema.findAll({
+    Request: ListProjectionThreadProposedPlansInput,
+    Result: ProjectionThreadProposedPlan,
+    execute: ({ threadId }) => sql`
+      SELECT
+        plan_id AS "planId",
+        thread_id AS "threadId",
+        turn_id AS "turnId",
+        plan_markdown AS "planMarkdown",
+        implemented_at AS "implementedAt",
+        implementation_thread_id AS "implementationThreadId",
+        created_at AS "createdAt",
+        updated_at AS "updatedAt"
+      FROM projection_thread_proposed_plans
+      WHERE thread_id = ${threadId}
+      ORDER BY created_at ASC, plan_id ASC
+    `,
+  });
+
+  const deleteProjectionThreadProposedPlanRows = SqlSchema.void({
+    Request: DeleteProjectionThreadProposedPlansInput,
+    execute: ({ threadId }) => sql`
+      DELETE FROM projection_thread_proposed_plans
+      WHERE thread_id = ${threadId}
+    `,
+  });
+
+  const upsert: ProjectionThreadProposedPlanRepository["Service"]["upsert"] = (row) =>
+    upsertProjectionThreadProposedPlanRow(row).pipe(
+      Effect.mapError(toPersistenceSqlError("ProjectionThreadProposedPlanRepository.upsert:query")),
+    );
+
+  const listByThreadId: ProjectionThreadProposedPlanRepository["Service"]["listByThreadId"] = (
+    input,
+  ) =>
+    listProjectionThreadProposedPlanRows(input).pipe(
+      Effect.mapError(
+        toPersistenceSqlError("ProjectionThreadProposedPlanRepository.listByThreadId:query"),
+      ),
+    );
+
+  const deleteByThreadId: ProjectionThreadProposedPlanRepository["Service"]["deleteByThreadId"] = (
+    input,
+  ) =>
+    deleteProjectionThreadProposedPlanRows(input).pipe(
+      Effect.mapError(
+        toPersistenceSqlError("ProjectionThreadProposedPlanRepository.deleteByThreadId:query"),
+      ),
+    );
+
+  return {
+    upsert,
+    listByThreadId,
+    deleteByThreadId,
+  } satisfies ProjectionThreadProposedPlanRepository["Service"];
+});
+
+export const layer = Layer.effect(ProjectionThreadProposedPlanRepository, make);
